belajar java dasar

+++ src/KonversiNumber.java
@@ -6,7 +6,7 @@ public class KonversiNumber {
         int iniInt = iniShort;
 
         int iniInt2 = 1000;
-        byte iniByte2 = iniInt2;
+        byte iniByte2 = (byte) iniInt2;
 
     }
 }
